move BookmarkFeatureControl from :preferences to :utils
